Extracting an abstract Configurator class for reuse.
